Fix overflow
git-svn-id: https://svn.code.sf.net/p/mtsieve/svn@344 bb85bfea-1d01-4b18-b6d8-510d50b5ba24

+++ twin/TwinApp.cpp
@@ -179,7 +179,7 @@ void TwinApp::ValidateOptions(void)
       if (it_TermType == TT_FACTORIAL)
          BuildFactorialTerms();
       
-      uint32_t termSize = (il_MaxK - il_MinK) + 1;
+      uint64_t termSize = (il_MaxK - il_MinK) + 1;
       
       if (ib_HalfK)
          termSize = (il_MaxK - il_MinK)/2 + 1;
